Generalize and move color error computation

+++ src/appleseed/renderer/modeling/aov/pixelerroraov.cpp
@@ -97,10 +97,9 @@ namespace
                     float ref_color[4];
                     frame.ref_image().get_pixel(x, y, ref_color);
 
-                    float error = static_cast<float>(
-                        sqrt(compute_error_squared(
-                            Color3f::from_array(image_color),
-                            Color3f::from_array(ref_color))));
+                    float error = sqrt(compute_error_squared(
+                        Color3f::from_array(image_color),
+                        Color3f::from_array(ref_color)));
                     max_error = max(max_error, error);
 
                     const Color3f color(error, 0, 0);
